move test_local_registry to it tests
